Simplify getting a HAXWindow for an NNWindow
HAXWindows were pretty flaky, with their elementRefs dying
unexpectedly. It's better to precache on object creation and let the
cache handle invalidation, since that's something it's really really
good at. At some point in the future it might be worth it to re-fetch
as soon as the object dies instead of on next request but this is good
for now.