Performance improvements
NNKit commit 6107640b7d051bb34431dcf01030815537937266 significantly
reduces dogpiling on the part of the window content polling objects,
but the real change here is moving `[self.store
startUpdatingWindowContents]` to run after the interface has been set
to display—the rendering thread takes long enough when the views its
rendering are not changing, updating the contents of the most expensive
layers before any have been drawn to the screen creates a feedback
loop: the renderer is delayed, having to catch up on more updates,
which gives the polling objects more time to queue more updates.

Unfortunately #30 is still a problem, but this improves things
significantly.