Fix #68.
Switch was using the relaunch mechanism built into Sparkle, which changed between the current shipping binary and the top of master, which Switch started using in 0.0.7. This ultimately meant I traded a corrupted code signature for a crash on the very first launch. Oops.

Switch now includes its own relaunch mechanism, which is a (minor) rewrite of Sparkle's old relauncher.